prefer symlinks over hardlinks
Btrfs is known to have a low limit on hardlinks to the same file in the same
directory but this depends on the names of the links. While we are falling back
to symlinks while building and installing, after that we don't have control
anymore. For example using ext4 as build/install fs but btrfs as an backup fs,
you will hit probably this limit when preserving hardlinks while backing up.
And if you are using ecryptfs, you will have very long file names. So let the
user decide whether he prefers symlinks over hardlinks, even if the
build/install fs would allow hardlinks.

Signed-off-by: Bert Wesarg <bert.wesarg@googlemail.com>